Remove tasks v1 endpoint and consolidate code to v2 files
