+++ backend/api/router.go
@@ -31,7 +31,6 @@ func GetRouter(handlers *API) *gin.Engine {
 	router.DELETE("/linked_accounts/:account_id/", handlers.DeleteLinkedAccount)
 	router.GET("/events/", handlers.EventsList)
 	router.GET("/messages/", handlers.MessagesList)
-	router.GET("/tasks/", handlers.TasksList)
 	router.GET("/tasks/v2/", handlers.TasksListV2)
 	router.POST("/tasks/create/:source_id/", handlers.TaskCreate)
 	router.PATCH("/tasks/modify/:task_id/", handlers.TaskModify)
